Updated the shield power up and scripts to work.
Currently the player is able to pick up a shield power up and when player is damaged, the shield will take damage first before the player's health. The player can only pick up another shield power up if they do not have one currently on. They cannot pick up another shield power up to refresh their current shield power up. For example, if their current shield health is at 50% they cannot pick up another shield to refresh it back to 100%. (Possibly might make this a game feature.)

+++ Project - Game Submission/Beta Version/Galaxy Defender - Beta v1/Assets/Scripts/PlayerHealth.cs
@@ -15,7 +15,7 @@ public class PlayerHealth : MonoBehaviour
         currentHealth = maxHealth;
     }
 
-    public void TakeDamage(int DamageTaken)
+    public void TakeDamage(float DamageTaken)
     {
         currentHealth -= DamageTaken;
